+++ src/sessionswindow/sessionswindow.h
@@ -22,7 +22,10 @@
 
 #include <QObject>
 #include <QWidget>
+#include <QTcpServer>
 #include <QTcpSocket>
+#include <QUdpSocket>
+#include <QNetworkDatagram>
 #include <QSerialPort>
 #include <QLocalSocket>
 #include <QMutex>
@@ -38,6 +41,194 @@
 #include "ptyqt.h"
 #include "qvncclientwidget.h"
 #include "qextserialenumerator.h"
+
+class QRawSocket : public QObject
+{
+    Q_OBJECT
+public:
+    QRawSocket(QObject *parent = nullptr)
+        : QObject(parent){}
+    ~QRawSocket() {
+        disconnectFromHost();
+        switch (m_rawMode) {
+        case 0:
+            delete tcpSocket;
+            break;
+        case 1:
+            if(tcpSocket) {
+                delete tcpSocket;
+            }
+            delete tcpServer;
+            break;
+        case 2:
+        case 3:
+            delete udpSocket;
+            break;
+        default:
+            break;
+        }
+    }
+    int setRawMode(int mode) {
+        if(m_rawMode == -1) {
+            switch (mode) {
+            case 0:
+                tcpSocket = new QTcpSocket(this);
+                connect(tcpSocket,&QTcpSocket::readyRead,this,&QRawSocket::readyRead);
+                connect(tcpSocket,&QTcpSocket::errorOccurred,this,&QRawSocket::errorOccurred);
+                connect(tcpSocket,&QTcpSocket::stateChanged,this,&QRawSocket::stateChanged);
+                break;
+            case 1:
+                tcpServer = new QTcpServer(this);
+                connect(tcpServer, &QTcpServer::newConnection, [&](){
+                    if (tcpSocket) {
+                        // If there is already a connection, reject the new one
+                        QTcpSocket *newSocket = tcpServer->nextPendingConnection();
+                        newSocket->disconnectFromHost();
+                        newSocket->deleteLater();
+                        return;
+                    }
+                    tcpSocket = tcpServer->nextPendingConnection();
+                    connect(tcpSocket,&QTcpSocket::readyRead,this,&QRawSocket::readyRead);
+                    connect(tcpSocket,&QTcpSocket::errorOccurred,this,&QRawSocket::errorOccurred);
+                    connect(tcpSocket,&QTcpSocket::stateChanged,this,&QRawSocket::stateChanged);
+                    emit stateChanged(tcpSocket->state());
+                });
+                break;
+            case 2:
+                udpSocket = new QUdpSocket(this);
+                connect(udpSocket,&QUdpSocket::errorOccurred,this,&QRawSocket::errorOccurred);
+                break;
+            case 3:
+                udpSocket = new QUdpSocket(this);
+                connect(udpSocket,&QUdpSocket::readyRead,this,&QRawSocket::readyRead);
+                connect(udpSocket,&QUdpSocket::errorOccurred,this,&QRawSocket::errorOccurred);
+                break;
+            default:
+                return -1;
+            }
+            m_rawMode = mode;
+            return 0;
+        }
+        return -1;
+    }
+    int getRawMode() const { return m_rawMode; }
+    void connectToHost(const QString &hostName, quint16 port) {
+        switch(m_rawMode) {
+        case 0:
+            tcpSocket->connectToHost(hostName,port);
+            return;
+        case 1:
+            tcpServer->listen(QHostAddress(hostName),port);
+            return;
+        case 2:
+            udpSocket->connectToHost(hostName,port);
+            udpstate = QAbstractSocket::ConnectedState;
+            emit stateChanged(udpstate);
+            return;
+        case 3:
+            udpSocket->bind(QHostAddress(hostName),port);
+            udpstate = QAbstractSocket::ConnectedState;
+            emit stateChanged(udpstate);
+            return;
+        }
+    }
+    void disconnectFromHost() {
+        switch(m_rawMode) {
+        case 0:
+            tcpSocket->disconnectFromHost();
+            return;
+        case 1:
+            if(tcpSocket)
+                tcpSocket->close();
+            tcpServer->close();
+            return;
+        case 2:
+            udpSocket->disconnectFromHost();
+            udpstate = QAbstractSocket::ClosingState;
+            emit stateChanged(udpstate);
+            return;
+        case 3:
+            udpSocket->close();
+            udpstate = QAbstractSocket::ClosingState;
+            emit stateChanged(udpstate);
+            return;
+        }
+    }
+    QByteArray readAll() {
+        switch(m_rawMode) {
+        case 0:
+            return tcpSocket->readAll();
+        case 1:
+            if(tcpSocket)
+                return tcpSocket->readAll();
+            return QByteArray();
+        case 3:
+            QByteArray data;
+            while (udpSocket->hasPendingDatagrams()) {
+                QNetworkDatagram datagram = udpSocket->receiveDatagram();
+                data.append(datagram.data());
+            }
+            return data;
+        }
+        return QByteArray();
+    }
+    qint64 write(const QByteArray &data) {
+        switch(m_rawMode) {
+        case 0:
+            return tcpSocket->write(data);
+        case 1:
+            if(tcpSocket)
+                return tcpSocket->write(data);
+            return 0;
+        case 2:
+            return udpSocket->write(data);
+        }
+        return 0;
+    }
+    qint64 write(const char *data, qint64 len) {
+        return this->write(QByteArray(data,len));
+    }
+    QAbstractSocket::SocketState state() const {
+        switch(m_rawMode) {
+        case 0:
+            return tcpSocket->state();
+        case 1:
+            if(tcpSocket)
+                return tcpSocket->state();
+            return QAbstractSocket::UnconnectedState;
+        case 2:
+        case 3:
+            return udpstate;
+        }
+        return QAbstractSocket::UnconnectedState;
+    }
+    QString errorString() const {
+        switch(m_rawMode) {
+        case 0:
+            return tcpSocket->errorString();
+        case 1:
+            if(tcpSocket)
+                return tcpSocket->errorString();
+            return QString();
+        case 2:
+        case 3:
+            return udpSocket->errorString();
+        }
+        return QString();
+    }
+
+signals:
+    void readyRead();
+    void errorOccurred(QAbstractSocket::SocketError socketError);
+    void stateChanged(QAbstractSocket::SocketState);
+
+private:
+    QTcpSocket *tcpSocket = nullptr;
+    QTcpServer *tcpServer = nullptr;
+    QUdpSocket *udpSocket = nullptr;
+    int m_rawMode = -1;
+    QAbstractSocket::SocketState udpstate = QAbstractSocket::ClosingState;
+};
 
 class SessionsWindow : public QObject
 {
@@ -133,7 +324,7 @@ public:
     int startTelnetSession(const QString &hostname, quint16 port, QTelnet::SocketType type);
     int startSerialSession(const QString &portName, uint32_t baudRate,
                     int dataBits, int parity, int stopBits, bool flowControl, bool xEnable );
-    int startRawSocketSession(const QString &hostname, quint16 port);
+    int startRawSocketSession(const QString &hostname, quint16 port, int mode);
     int startNamePipeSession(const QString &name);
 #ifdef ENABLE_SSH
     int startSSH2Session(const QString &hostname, quint16 port, const QString &username, const QString &password);
@@ -234,6 +425,7 @@ public:
     quint16 getPort() const { return m_port; }
     QString getPortName() const { return m_portName; }
     QTelnet::SocketType getSocketType() const { return m_type; }
+    int getRawMode() const { if(rawSocket) return rawSocket->getRawMode(); else return 0;}
     uint32_t getBaudRate() const { return m_baudRate; }
     int getDataBits() const { return m_dataBits; }
     int getParity() const { return m_parity; }
@@ -474,7 +666,7 @@ private:
     QTelnet *telnet;
     QSerialPort *serialPort;
     QextSerialEnumerator *serialMonitor;
-    QTcpSocket *rawSocket;
+    QRawSocket *rawSocket;
     IPtyProcess *localShell;
     QLocalSocket *namePipe;
 #ifdef ENABLE_SSH